+++ example/src/components/index.js
@@ -1,0 +1,7 @@
+export { default as CseView } from './CseView';
+export { default as SettingView } from './SettingsView';
+export { default as ResultView } from './ResultView';
+export { default as SessionsCheckout } from './Checkout/SessionsCheckout';
+export { default as AdvancedCheckout } from './Checkout/AdvancedCheckout';
+export { default as Home } from './HomeView';
+export { default as PartialPaymentCheckout } from './Checkout/PartialPaymentCheckout';
