+++ example/src/components/utilities/isSuccess.ts
@@ -1,9 +1,9 @@
 import { ResultCode } from '@adyen/react-native';
 
-export const isSuccess = ({ resultCode }) =>
+export const isSuccess = (code: ResultCode) =>
   [
     ResultCode.authorised,
     ResultCode.received,
     ResultCode.pending,
     ResultCode.presentToShopper,
-  ].includes(resultCode);
+  ].includes(code);
