+++ example/src/components/utilities/payWithCard.ts
@@ -1,14 +1,15 @@
-import { AdyenCSE, AdyenAction } from '@adyen/react-native';
-import { ENVIRONMENT } from '../Configuration';
-import ApiClient from '../api/APIClient';
-import { checkoutConfiguration } from '../State/checkoutConfiguration';
+import { AdyenCSE, AdyenAction, Card, type PaymentMethodData } from '@adyen/react-native';
+import { ENVIRONMENT } from '../../Configuration';
+import ApiClient from '../../api/APIClient';
+import { checkoutConfiguration } from '../../State/checkoutConfiguration';
+import type { PaymentConfiguration } from '../../api/types';
 
-export async function payWithCard(unencryptedCard, configuration) {
+export async function payWithCard(unencryptedCard: Card, configuration: PaymentConfiguration) {
   const encryptedCard = await AdyenCSE.encryptCard(
     unencryptedCard,
     ENVIRONMENT.publicKey
   );
-  const paymentData = {
+  const paymentData: PaymentMethodData = {
     paymentMethod: {
       type: 'scheme',
       encryptedCardNumber: encryptedCard.number,
@@ -17,12 +18,12 @@ export async function payWithCard(unencryptedCard, configuration) {
       encryptedSecurityCode: encryptedCard.cvv,
       threeDS2SdkVersion: AdyenAction.threeDS2SdkVersion,
     },
+    returnUrl: ENVIRONMENT.returnUrl
   };
 
   let result = await ApiClient.payments(
     paymentData,
-    configuration,
-    ENVIRONMENT.returnUrl
+    configuration
   );
   if (result.action) {
     const actionConfiguration = checkoutConfiguration(configuration);
